+++ nginx-nomad-alt/CHANGELOG.md
@@ -3,6 +3,7 @@
 * First release nginx-nomad-alt flavour
 * This provides a services-based nginx and php nomad job
 * Fixup nginx.conf and create /mnt/www if necessary
+* Revert to standard /usr/local/www/stdwebsite and fix docs and nomad job
 
 ---